tests/requirements/dj60_cms50.txt
-r base.txt

Django>=6.0,<6.1
django-cms>=5.0,<5.1
djangocms-versioning>=2.4
